# domain/user.go
package domain

import (
	"context"
	"errors"
	"time"
)

var (
	ErrInvalidPassword   = errors.New("invalid password")
	ErrCannotDeleteSelf  = errors.New("不能删除自己")
	ErrCannotDeleteAdmin = errors.New("不能删除管理员账户")
	ErrCannotEditAdmin   = errors.New("不能编辑管理员账户")
	ErrUserDisabled      = errors.New("账户已停用或未启用")
)

// User 结构体定义了用户的基本信息 (单租户架构)
type User struct {
	ID             string     `json:"id"`
	Username       string     `json:"username"`
	Nickname       string     `json:"nickname"`
	Email          string     `json:"email"`
	Phone          string     `json:"phone,omitempty"`
	Password       string     `json:"password,omitempty"`
	Source         string     `json:"source,omitempty"` // local-本地用户，oauth-第三方来源用户
	Avatar         string     `json:"avatar,omitempty"`
	IsAdmin        bool       `json:"is_admin,omitempty"`
	Status         string     `json:"status"`              // active, inactive, invited, suspended
	Roles          []string   `json:"role,omitempty"`      // 用户的角色
	IsActive       bool       `json:"is_active,omitempty"` // 用户是否活跃
	DepartmentID   *string    `json:"department_id,omitempty"`
	DepartmentName string     `json:"department_name,omitempty"`
	CreatedAt      time.Time  `json:"created_at"`
	UpdatedAt      time.Time  `json:"updated_at"`
	InvitedAt      *time.Time `json:"invited_at,omitempty"`
	InvitedBy      string     `json:"invited_by,omitempty"`
}

// CreateUserRequest 创建用户请求 (单租户架构)
type CreateUserRequest struct {
	Username     string   `json:"username" binding:"required"`
	Email        string   `json:"email" binding:"required,email"`
	Phone        string   `json:"phone,omitempty"`
	Password     string   `json:"password" binding:"required"`
	Status       string   `json:"status,omitempty"`        // active, inactive, invited, suspended
	RoleIDs      []string `json:"role_ids,omitempty"`      // 用户角色ID列表
	DepartmentID *string  `json:"department_id,omitempty"` // 所属部门ID
}

// InviteUserRequest 邀请用户请求 (单租户架构)
type InviteUserRequest struct {
	Email   string   `json:"email" binding:"required,email"`
	RoleIDs []string `json:"role_ids,omitempty"` // 角色ID列表
	Message string   `json:"message,omitempty"`  // 可选的邀请消息
}

// UserUpdateRequest 更新用户请求
type UserUpdateRequest struct {
	Username     *string  `json:"username,omitempty"`
	Email        *string  `json:"email,omitempty"`
	Phone        *string  `json:"phone,omitempty"`
	Password     *string  `json:"password,omitempty"`
	Status       *string  `json:"status,omitempty"`
	Avatar       *string  `json:"avatar,omitempty"`
	RoleIDs      []string `json:"role_ids,omitempty"`
	DepartmentID *string  `json:"department_id,omitempty"`
}

// UserPagedResult 用户分页查询结果
type UserPagedResult = PagedResult[*User]

// UserQueryFilter 用户查询过滤器 (单租户架构)
type UserQueryFilter struct {
	Status       string `json:"status,omitempty"`
	Role         string `json:"role,omitempty"`
	Username     string `json:"username,omitempty"`
	Email        string `json:"email,omitempty"`
	Keyword      string `json:"keyword,omitempty"` // 模糊匹配 username / nickname / email（CLI 统一约定）
	IsAdmin      *bool  `json:"is_admin,omitempty"`
	IncludeRoles bool   `json:"include_roles,omitempty"`
	DepartmentID string `json:"department_id,omitempty" form:"department_id"`
	QueryParams
}

type UserRepository interface {
	Create(c context.Context, user *User) error
	GetByID(c context.Context, id string) (*User, error)
	GetByIdentifier(c context.Context, identifier string) (*User, error)
	Query(c context.Context, filter UserQueryFilter) (*PagedResult[*User], error)
	Update(c context.Context, user *User) error
	Delete(c context.Context, id string) error
	// GetStatusByID returns only the status string for the user.
	GetStatusByID(c context.Context, id string) (string, error)
	GetRoleIDs(c context.Context, id string) ([]string, error)
}

type UserUseCase interface {
	CreateUser(c context.Context, request *CreateUserRequest) (*User, error)
	GetUserByID(c context.Context, id string) (*User, error)
	ListUsers(c context.Context, filter UserQueryFilter) (*PagedResult[*User], error)
	UpdateUserProfile(c context.Context, userID string, updates ProfileUpdate) error
	UpdateUserPassword(c context.Context, userID string, passwordUpdate PasswordUpdate) error
	UpdateUserPartial(c context.Context, userID string, updates UserUpdateRequest) error
	DeleteUser(c context.Context, id string, currentUserID string) error
	InviteUser(c context.Context, request *InviteUserRequest, invitedBy string) (*User, error)
	GetUserRoles(c context.Context, userID string) ([]string, error)
}